tf_tutorial/requirements.txt
tf-nightly
matplotlib

torchtext
transformers
